% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDT_Fcns.R
\name{BDT_Calibration}
\alias{BDT_Calibration}
\title{BDT_Calibration}
\usage{
BDT_Calibration(.b, .short_mkt)
}
\arguments{
\item{.b}{- b0 parameter of the BDT model}

\item{.short_mkt}{- yield curve}
}
\value{
A list with the calibrated parameter, the sum of squared residuals of the difference, the short run lattice and the element prices
}
\description{
Calibrate the a parameters of the BDT model from the observed Yield Curve
}
