% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDT_Fcns.R
\name{Elem_Prices}
\alias{Elem_Prices}
\title{Elem_Prices}
\usage{
Elem_Prices(.R)
}
\arguments{
\item{.R}{A Lattice for the Short Run Rate}
}
\value{
A Lattice for the Element Prices
}
\description{
Returns a lattice of Element Prices (Contingent Claim Prices) from a Short Run Interest Lattice
}
