% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events_BRbond.R
\name{Events_BRbond}
\alias{Events_BRbond}
\title{Function for generating a Table of Events of Brazilian Debt Instruments}
\usage{
Events_BRbond(dataEval = NULL, maturity, coupon_rate = NULL, VF, YTM,
  months_coupon)
}
\arguments{
\item{dataEval}{Date in which the bond is to be priced. If NULL, the price is for the current date.}

\item{maturity}{Maturity date.}

\item{coupon_rate}{Coupon Rate. When NULL, it returns the price of a zero coupon.}

\item{VF}{Face value of the bond.}

\item{YTM}{Yield to Maturity.}

\item{months_coupon}{Number of months between each coupon. If coupon is NULL, not used.}
}
\value{
Bond Price.
}
\description{
Function for generating a Table of Events of Brazilian Debt Instruments
}
