% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events_LTN_BR.R
\name{Events_LTN_BR}
\alias{Events_LTN_BR}
\title{Function for Getting Event List for ``Letras do Tesouro Nacional'' 
Zero Coupon Brazilian government bonds, also known as ``Tesouro Prefixado''}
\usage{
Events_LTN_BR(.dataEval = NULL, .maturity, .YTM)
}
\arguments{
\item{.dataEval}{- Date when the pricing is to be made. If NULL, it is priced in the current working day}

\item{.maturity}{- Maturity - as a string or a date (Date or POSIXct)}

\item{.YTM}{- Yield to Maturity (in decimals per year with 252 working days)}
}
\value{
Data Frame with dates (in days), Cash Flows and PV of Cash Flows
}
\description{
Function for Getting Event List for ``Letras do Tesouro Nacional'' 
Zero Coupon Brazilian government bonds, also known as ``Tesouro Prefixado''
}
