% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PChg_BRbond.R
\name{PChg_BRbond}
\alias{PChg_BRbond}
\title{Function for estimating the predicted price change for a Brazilian bond}
\usage{
PChg_BRbond(dataEval = NULL, maturity, coupon_rate = NULL, VF, YTM,
  months_coupon, dYTM)
}
\arguments{
\item{dataEval}{Date in which the bond is to be priced. If NULL, the price is for the current date.}

\item{maturity}{Maturity date.}

\item{coupon_rate}{Coupon Rate. When NULL, it returns the price of a zero coupon.}

\item{VF}{Face value of the bond.}

\item{YTM}{Yield to Maturity.}

\item{months_coupon}{Number of months between each coupon. If coupon is NULL, not used.}

\item{dYTM}{Predicted change in YTM (in basis points)}
}
\value{
Predicted percentage change in bond price
}
\description{
Function for estimating the predicted price change for a Brazilian bond
}
