% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PU_LTN_BDT.R
\name{PU_LTN_BDT}
\alias{PU_LTN_BDT}
\title{PU_LTN_BDT}
\usage{
PU_LTN_BDT(.dataEval = NULL, .maturity, .nstates, .beta, .lambda)
}
\arguments{
\item{.dataEval}{- Data as which the price is to be calculated}

\item{.maturity}{- Maturity}

\item{.nstates}{- Number of time periods in the lattice}

\item{.beta}{- Vector of Beta coefficients for the ANBIMA model of the term structure}

\item{.lambda}{- Vector of Lambda coefficients for the ANBUMA model of the term structure}
}
\value{
- Price in BRL of a LTN
}
\description{
Function which generates the PU of a LTN using the Black-Derman Toy Model
}
