% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PU_NTNF_BR.R
\name{PU_NTNF_BR}
\alias{PU_NTNF_BR}
\title{Function for pricing ``Notas do Tesouro Nacional - Série F'' 
Zero Coupon Brazilian government bonds, also known as ``Tesouro Prefixado com Juros Semestrais''}
\usage{
PU_NTNF_BR(.dataEval = NULL, .maturity, .YTM)
}
\arguments{
\item{.dataEval}{- Date when the pricing is to be made. If NULL, it is priced in the current working day}

\item{.maturity}{- Maturity - as a string or a date (Date or POSIXct)}

\item{.YTM}{- Yield to Maturity (in decimals per year with 252 working days)}
}
\value{
Bond Price
}
\description{
Function for pricing ``Notas do Tesouro Nacional - Série F'' 
Zero Coupon Brazilian government bonds, also known as ``Tesouro Prefixado com Juros Semestrais''
}
