% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDT_Fcns.R
\name{SR_Lattice_BDT}
\alias{SR_Lattice_BDT}
\title{SR_Lattice_BDT}
\usage{
SR_Lattice_BDT(.b, .a)
}
\arguments{
\item{.b}{- b0 parameter of the BDT Model}

\item{.a}{- vector of a parameter of the model}
}
\value{
- Short-Run Lattice of interest rates
}
\description{
Construct a Lattice for the Short Run Rate using the Black-Derman-Toy Model.
}
