% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yield_Curve.R
\name{Yield_Curve}
\alias{Yield_Curve}
\title{Yield Curve}
\usage{
Yield_Curve(.beta = c(0.113525690989684, -0.0463197432094039,
  -0.0709736348702071, -0.00365241403584365), .lambda = c(1.61175546741901,
  0.428872702898983), .t)
}
\arguments{
\item{.beta}{- Vector of beta coefficients}

\item{.lambda}{- Vector of lambda coefficients}

\item{.t}{- Time in years of 252 working days}
}
\value{
- The YTM (in decimals) per year of 252 working days
}
\description{
Estimates the YTM of a bond maturing in .t periods according to the ANBIMA model of the Brazilian Yield Curve.
Defaults for .beta and .lambda are values of May 18th, 2015. To update, get the most recent values at http://www.anbima.com.br/est_termo/CZ.asp
}
