% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YTM_BRbond.R
\name{YTM_BRbond}
\alias{YTM_BRbond}
\title{Function for finding the YTM for Brazilian Debt Instruments}
\usage{
YTM_BRbond(dataEval = NULL, maturity, coupon_rate = NULL, VF, PU,
  months_coupon)
}
\arguments{
\item{dataEval}{Date in which the bond is to be priced. If NULL, the price is for the current date.}

\item{maturity}{Maturity date.}

\item{coupon_rate}{Coupon Rate. When NULL, it returns the price of a zero coupon.}

\item{VF}{Face value of the bond.}

\item{PU}{PU.}

\item{months_coupon}{Number of months between each coupon. If coupon is NULL, not used.}
}
\value{
Bond Price.
}
\description{
Function for finding the YTM for Brazilian Debt Instruments
}
