% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PChg_BRbond.R
\name{PChg_BRbond}
\alias{PChg_BRbond}
\title{PChg_BRbond
Function to compute percentage change in Brazilian bond prices given a change in Yield to Maturity 
Note: it only computes changes for the current date}
\usage{
PChg_BRbond(type, .maturity, .YTM, .dYTM)
}
\arguments{
\item{type}{String specifying which bond type. Currently accepted "NTNF", "NTNB", "NTNBc" and "LTN"}

\item{.maturity}{Maturity of the Bond}

\item{.YTM}{Current Yield to Maturity}

\item{.dYTM}{Proposed change}
}
\description{
PChg_BRbond
Function to compute percentage change in Brazilian bond prices given a change in Yield to Maturity 
Note: it only computes changes for the current date
}
